services/file-processing-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
